move stuff from binary to pg
