+++ squeal-postgresql/src/Squeal/PostgreSQL/Definition.hs
@@ -96,7 +96,6 @@ import qualified Generics.SOP as SOP
 import qualified GHC.Generics as GHC
 
 import Squeal.PostgreSQL.Alias
-import Squeal.PostgreSQL.Binary
 import Squeal.PostgreSQL.Expression
 import Squeal.PostgreSQL.PG
 import Squeal.PostgreSQL.Query
